align hierarchy.py 1

+++ director/hierarchy.py
@@ -90,10 +90,10 @@ class Hierarchy(tfutils.Module):
     skill = sg(switch(carry['skill'], self.manager.actor(sg(latent)).sample()))
     new_goal = self.dec({'skill': skill, 'context': self.feat(latent)}).mode()
     new_goal = (
-        self.feat(latent).astype(tf.float32) + new_goal
+        self.feat(latent) + new_goal
         if self.config.manager_delta else new_goal)
     goal = sg(switch(carry['goal'], new_goal))
-    delta = goal - self.feat(latent).astype(tf.float32)
+    delta = goal - self.feat(latent)
     dist = self.worker.actor(sg({**latent, 'goal': goal, 'delta': delta}))
     outs = {'action': dist}
     if 'image' in self.wm.heads['decoder'].shapes:
@@ -104,7 +104,7 @@ class Hierarchy(tfutils.Module):
     return outs, carry
 
   def train(self, imagine, start, data):
-    success = lambda rew: (rew[-1] > 0.7).astype(tf.float32).mean()
+    success = lambda rew: (rew[-1] > 0.7).mean()
     metrics = {}
     if self.config.expl_rew == 'disag':
       metrics.update(self.expl_reward.train(data))
@@ -155,7 +155,7 @@ class Hierarchy(tfutils.Module):
       traj['reward_extr'] = self.extr_reward(traj)
       traj['reward_expl'] = self.expl_reward(traj)
       traj['reward_goal'] = self.goal_reward(traj)
-      traj['delta'] = traj['goal'] - self.feat(traj).astype(tf.float32)
+      traj['delta'] = traj['goal'] - self.feat(traj)
       wtraj = self.split_traj(traj)
       mtraj = self.abstract_traj(traj)
     mets = self.worker.update(wtraj, tape)
@@ -173,7 +173,7 @@ class Hierarchy(tfutils.Module):
       skill = self.manager.actor(sg(start)).sample()
       goal = self.dec({'skill': skill, 'context': context}).mode()
       goal = (
-          self.feat(start).astype(tf.float32) + goal
+          self.feat(start) + goal
           if self.config.manager_delta else goal)
       worker = lambda s: self.worker.actor(sg({
           **s, 'goal': goal, 'delta': goal - self.feat(s)})).sample()
@@ -183,7 +183,7 @@ class Hierarchy(tfutils.Module):
       traj['reward_extr'] = self.extr_reward(traj)
       traj['reward_expl'] = self.expl_reward(traj)
       traj['reward_goal'] = self.goal_reward(traj)
-      traj['delta'] = traj['goal'] - self.feat(traj).astype(tf.float32)
+      traj['delta'] = traj['goal'] - self.feat(traj)
       wtraj = traj.copy()
       mtraj = self.abstract_traj_old(traj)
     mets = self.worker.update(wtraj, tape)
@@ -202,7 +202,7 @@ class Hierarchy(tfutils.Module):
       traj['reward_extr'] = self.extr_reward(traj)
       traj['reward_expl'] = self.expl_reward(traj)
       traj['reward_goal'] = self.goal_reward(traj)
-      traj['delta'] = traj['goal'] - self.feat(traj).astype(tf.float32)
+      traj['delta'] = traj['goal'] - self.feat(traj)
       mtraj = self.abstract_traj(traj)
     metrics = self.manager.update(mtraj, tape)
     metrics = {f'manager_{k}': v for k, v in metrics.items()}
@@ -214,21 +214,21 @@ class Hierarchy(tfutils.Module):
     sg = lambda x: tf.nest.map_structure(tf.stop_gradient, x)
     with tf.GradientTape(persistent=True) as tape:
       worker = lambda s: self.worker.actor(sg({
-          **s, 'goal': goal, 'delta': goal - self.feat(s).astype(tf.float32),
+          **s, 'goal': goal, 'delta': goal - self.feat(s),
       })).sample()
       traj = imagine(worker, start, self.config.imag_horizon)
       traj['goal'] = tf.repeat(goal[None], 1 + self.config.imag_horizon, 0)
       traj['reward_extr'] = self.extr_reward(traj)
       traj['reward_expl'] = self.expl_reward(traj)
       traj['reward_goal'] = self.goal_reward(traj)
-      traj['delta'] = traj['goal'] - self.feat(traj).astype(tf.float32)
+      traj['delta'] = traj['goal'] - self.feat(traj)
     mets = self.worker.update(traj, tape)
     metrics.update({f'worker_{k}': v for k, v in mets.items()})
     return traj, metrics
 
   def train_vae_replay(self, data):
     metrics = {}
-    feat = self.feat(data).astype(tf.float32)
+    feat = self.feat(data)
     if 'context' in self.config.goal_decoder.inputs:
       if self.config.vae_span:
         context = feat[:, 0]
@@ -258,7 +258,7 @@ class Hierarchy(tfutils.Module):
 
   def train_vae_imag(self, traj):
     metrics = {}
-    feat = self.feat(traj).astype(tf.float32)
+    feat = self.feat(traj)
     if 'context' in self.config.goal_decoder.inputs:
       if self.config.vae_span:
         context = feat[0]
@@ -272,7 +272,7 @@ class Hierarchy(tfutils.Module):
     with tf.GradientTape() as tape:
       enc = self.enc({'goal': goal, 'context': context})
       dec = self.dec({'skill': enc.sample(), 'context': context})
-      rec = -dec.log_prob(tf.stop_gradient(goal.astype(tf.float32)))
+      rec = -dec.log_prob(tf.stop_gradient(goal))
       if self.config.goal_kl:
         kl = tfd.kl_divergence(enc, self.prior)
         kl, mets = self.kl(kl)
@@ -286,13 +286,13 @@ class Hierarchy(tfutils.Module):
     return metrics
 
   def propose_goal(self, start, impl):
-    feat = self.feat(start).astype(tf.float32)
+    feat = self.feat(start)
     if impl == 'replay':
-      target = tf.random.shuffle(feat).astype(tf.float32)
+      target = tf.random.shuffle(feat)
       skill = self.enc({'goal': target, 'context': feat}).sample()
       return self.dec({'skill': skill, 'context': feat}).mode()
     if impl == 'replay_direct':
-      return tf.random.shuffle(feat).astype(tf.float32)
+      return tf.random.shuffle(feat)
     if impl == 'manager':
       skill = self.manager.actor(start).sample()
       goal = self.dec({'skill': skill, 'context': feat}).mode()
@@ -304,9 +304,9 @@ class Hierarchy(tfutils.Module):
     raise NotImplementedError(impl)
 
   def goal_reward(self, traj):
-    feat = self.feat(traj).astype(tf.float32)
-    goal = tf.stop_gradient(traj['goal'].astype(tf.float32))
-    skill = tf.stop_gradient(traj['skill'].astype(tf.float32))
+    feat = self.feat(traj)
+    goal = tf.stop_gradient(traj['goal'])
+    skill = tf.stop_gradient(traj['skill'])
     context = tf.stop_gradient(
         tf.repeat(feat[0][None], 1 + self.config.imag_horizon, 0))
     if self.config.goal_reward == 'dot':
@@ -375,7 +375,7 @@ class Hierarchy(tfutils.Module):
     elif self.config.goal_reward == 'squared':
       return -((goal - feat) ** 2).sum(-1)[1:]
     elif self.config.goal_reward == 'epsilon':
-      return ((goal - feat).mean(-1) < 1e-3).astype(tf.float32)[1:]
+      return ((goal - feat).mean(-1) < 1e-3)[1:]
     elif self.config.goal_reward == 'enclogprob':
       return self.enc({'goal': goal, 'context': context}).log_prob(skill)[1:]
     elif self.config.goal_reward == 'encprob':
@@ -394,7 +394,7 @@ class Hierarchy(tfutils.Module):
       raise NotImplementedError(self.config.goal_reward)
 
   def elbo_reward(self, traj):
-    feat = self.feat(traj).astype(tf.float32)
+    feat = self.feat(traj)
     context = tf.repeat(feat[0][None], 1 + self.config.imag_horizon, 0)
     enc = self.enc({'goal': feat, 'context': context})
     dec = self.dec({'skill': enc.sample(), 'context': context})
@@ -478,7 +478,7 @@ class Hierarchy(tfutils.Module):
     goal = self.propose_goal(start, impl)
     # Worker rollout.
     worker = lambda s: self.worker.actor({
-        **s, 'goal': goal, 'delta': goal - self.feat(s).astype(tf.float32),
+        **s, 'goal': goal, 'delta': goal - self.feat(s),
     }).sample()
     traj = self.wm.imagine(
         worker, start, self.config.worker_report_horizon)
